+++ app/include/api/http.hpp
@@ -21,6 +21,10 @@ public:
     using Cancel = std::shared_ptr<std::atomic_bool>;
     using Progress = brls::Event<curl_off_t, curl_off_t>;
 
+    inline static bool PROXY_STATUS = false;
+    inline static std::string PROXY_HOST = "192.168.1.1";
+    inline static int PROXY_PORT = 1080;
+
     struct Range {
         int start = 0;
         int end = 0;